+++ Marlin/src/module/ft_motion/shaping.cpp
@@ -142,7 +142,7 @@ void AxisShaping::set_axis_shaping_A(
 // Ai[] must be precomputed (if zeta or vtol change, call set_axis_shaping_A first)
 void AxisShaping::set_axis_shaping_N(const ftMotionShaper_t shaper, const float f, const float zeta) {
   // Note that protections are omitted for DBZ and for index exceeding array length.
-  const float df = sqrt ( 1.f - sq(zeta) );
+  const float df = sqrtf ( 1.f - sq(zeta) );
   switch (shaper) {
     case ftMotionShaper_ZV:
       Ni[1] = LROUND((0.5f / f / df) * (FTM_FS));
